requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-admin-interface==0.18.2
django-allauth==0.46.0
django-braces==1.15.0
django-cleanup==5.2.0
django-colorfield==0.4.5
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-email-verification==0.1.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-model-utils==4.2.0
django-rest-auth==0.9.5
django-star-ratings==0.9.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.0.12
Flask==2.0.2
gunicorn==20.1.0
httpie==2.5.0
idna==3.2
isort==5.7.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.5.2
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.21
Pygments==2.8.1
PyJWT==2.3.0
pylint==2.7.0
PySocks==1.7.1
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
somepackage==1.2.3
sqlparse==0.4.1
swapper==1.2.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
